# Project/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  contact:
    email: "danielmcoelho@ua.pt"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
  /account:
    post:
      tags:
      - "account"
      summary: "Create an account with a uuid user"
      description: ""
      operationId: "create_account"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "formData"
        description: "ID of the user who will be associated with the created account"
        required: true
        type: "string"
        format: "uuid"
      - name: "currency"
        in: "formData"
        description: "The type of currency your account will be"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Account"
        400:
          description: 'Error code and reason to dispatch the error'
          schema:
            $ref: "#/definitions/ApiResponse"
  /account/{id}/amount:
    post:
      tags:
      - "account"
      summary: "Add an amount of money to an account"
      operationId: "add_amount"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
      - name : "id"
        in : "path"
        required: true
        type: "string"
        format: "uuid"
        description: "The identifier number of the account"
      - name: "amount"
        in : "formData"
        required: true
        type: "number"
        format: "float"
        description: "The amount added to account"          
      responses:
        200:
          description: "successful operation"
          schema:
            $ref : "#/definitions/ApiResponse"
        400:
          description: "something wrong with the arguments"
          schema:
            $ref : "#/definitions/ApiResponse"
        405:
          description: "method not allowed"
          schema:
            $ref : "#/definitions/ApiResponse"
  /account/{id}/activate:
    post:
      tags:
      - "account"
      summary: ""
      operationId: "activate_account"
      produces:
      - "application/json"
      parameters:
      - name : "id"
        in : "path"
        required: true
        type: "string"
        format: "uuid"
        description: "The identifier number of the account"
      responses:
          200:
            description: "successful operation"
            schema:
              $ref : "#/definitions/ApiResponse"
          304:
            description: "not modified"
            schema:
              $ref : "#/definitions/ApiResponse"
...
  /account/{id}:
    get:
      tags:
      - "account"
      summary: "Account Information"
      operationId: "account_info"
      produces:
      - "application/json"
      parameters:
      - name : "id"
        in : "path"
        required: true
        type: "string"
        format: "uuid"
        description: "The identifier number of the account"
      responses:
        200:
          description: "successful operation"
          schema:
            items:
              $ref: "#/definitions/Account"
        400:
          description: "Invalid ID supplied"
          schema:
            items:
              $ref: "#/definitions/ApiResponse"          
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Account:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
        description: "The account identifier number"
      user_id:
        type: "string"
        format: "uuid"
        description: "An user id identifier"        
      balance:
        type: "number"
        format: "float"
        description: "The actual amount in account"
      currency:
        type: "string"
        example: "EUR"
        description: "International standard that defines three-letter codes as currencies established by the International Organization. (ISO 4217)"
      state:
        type: "boolean"
        description: "The account state : ative or inactive"
      created_at:
        type: "string"
        format: "date-time"
        description: "The date when was created the account"
      updated_at:
        type: "string"
        format: "date-time"
        description: "The date when was updated the account"
    xml:
      name: "Account"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      response:
        type: "string"
        example : "reason of success or failure"